Extract admission registry into admission/registry.rs
